Move TrnRequestSupportTaskService's methods onto TrnRequestService
TrnRequestService was the only caller of most of TrnRequestSupportTaskService
and already took it as a dependency, so the split bought us nothing. Move
its four methods over and remove the class; TrnRequestService now depends on
SupportTaskService directly.

The options records move alongside them into Services/TrnRequests, leaving
Services/SupportTasks/TrnRequests empty, and the tests become a
TrnRequestServiceTests partial in the same vein as the matching ones.

AddSupportTaskService no longer has to register a second service for
TrnRequestService's benefit - it still supplies the SupportTaskService that
TrnRequestService needs.
