+++ src/TeachingRecordSystem.Core/Services/TrnRequests/CompleteManualChecksNeededSupportTaskOptions.cs
@@ -1,4 +1,4 @@
-namespace TeachingRecordSystem.Core.Services.SupportTasks.TrnRequests;
+namespace TeachingRecordSystem.Core.Services.TrnRequests;
 
 public record CompleteManualChecksNeededSupportTaskOptions
 {
