+++ src/TeachingRecordSystem.SupportUi/Pages/SupportTasks/TrnRequests/Resolve/CheckAnswers.cshtml.cs
@@ -3,7 +3,6 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
 using TeachingRecordSystem.Core.DataStore.Postgres;
 using TeachingRecordSystem.Core.Models.SupportTasks;
-using TeachingRecordSystem.Core.Services.SupportTasks.TrnRequests;
 using TeachingRecordSystem.Core.Services.TrnRequests;
 using TeachingRecordSystem.SupportUi.Services;
 using static TeachingRecordSystem.SupportUi.Pages.SupportTasks.TrnRequests.Resolve.ResolveTrnRequestState;
@@ -14,7 +13,6 @@ namespace TeachingRecordSystem.SupportUi.Pages.SupportTasks.TrnRequests.Resolve;
 public class CheckAnswers(
     TrsDbContext dbContext,
     TrnRequestService trnRequestService,
-    TrnRequestSupportTaskService trnRequestSupportTaskService,
     SupportUiLinkGenerator linkGenerator,
     TimeProvider timeProvider,
     PersonChangeableAttributesService changedService) :
@@ -105,7 +103,7 @@ public class CheckAnswers(
 
         var resolvedPersonAttributes = GetResolvedPersonAttributes(selectedPersonAttributes);
 
-        await trnRequestSupportTaskService.ResolveTrnRequestSupportTaskAsync(
+        await trnRequestService.ResolveTrnRequestSupportTaskAsync(
             new()
             {
                 SupportTaskReference = supportTask.SupportTaskReference,
